pyproject.toml
[project]
name = "aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.7",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "ruff>=0.15.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
